Refactor crypto module
